@@ -1,5 +1,15 @@
 <?php
 
+/*
+ You may not change or alter any portion of this comment or credits
+ of supporting developers from this source code or any supporting source code
+ which is considered copyrighted (c) material of the original comment or credit authors.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ */
+
 declare(strict_types=1);
 
 namespace Xoops\RegDom;
@@ -103,7 +113,7 @@ class RegisteredDomain
 
     private static function normalizeHost(string $input): string
     {
-        $host = (strpos($input, '/') !== false) ? parse_url($input, PHP_URL_HOST) : $input;
+        $host = str_contains($input, '/') ? parse_url($input, PHP_URL_HOST) : $input;
         if (!is_string($host)) {
             $host = '';
         }
