@@ -1,7 +1,7 @@
 <?php
 /**
  * Public Suffix List Cache
- * Generated: 2025-09-30 22:04:28 UTC
+ * Generated: 2026-02-07 09:24:29 UTC
  * Source: https://publicsuffix.org/list/public_suffix_list.dat
  * License: Mozilla Public License 2.0 (https://publicsuffix.org/)
  */
@@ -255,6 +255,21 @@ return array (
     'org.bb' => true,
     'store.bb' => true,
     'tv.bb' => true,
+    'bd' => true,
+    'ac.bd' => true,
+    'ai.bd' => true,
+    'co.bd' => true,
+    'com.bd' => true,
+    'edu.bd' => true,
+    'gov.bd' => true,
+    'id.bd' => true,
+    'info.bd' => true,
+    'it.bd' => true,
+    'mil.bd' => true,
+    'net.bd' => true,
+    'org.bd' => true,
+    'sch.bd' => true,
+    'tv.bd' => true,
     'be' => true,
     'ac.be' => true,
     'bf' => true,
@@ -871,7 +886,9 @@ return array (
     'ac.fj' => true,
     'biz.fj' => true,
     'com.fj' => true,
+    'edu.fj' => true,
     'gov.fj' => true,
+    'id.fj' => true,
     'info.fj' => true,
     'mil.fj' => true,
     'name.fj' => true,
@@ -1082,6 +1099,7 @@ return array (
     'ponpes.id' => true,
     'sch.id' => true,
     'web.id' => true,
+    'ᬩᬮᬶ.id' => true,
     'ie' => true,
     'gov.ie' => true,
     'il' => true,
@@ -1114,6 +1132,7 @@ return array (
     'ac.in' => true,
     'ai.in' => true,
     'am.in' => true,
+    'bank.in' => true,
     'bihar.in' => true,
     'biz.in' => true,
     'business.in' => true,
@@ -1127,6 +1146,7 @@ return array (
     'dr.in' => true,
     'edu.in' => true,
     'er.in' => true,
+    'fin.in' => true,
     'firm.in' => true,
     'gen.in' => true,
     'gov.in' => true,
@@ -3949,8 +3969,6 @@ return array (
     'bievát.no' => true,
     'bindal.no' => true,
     'birkenes.no' => true,
-    'bjarkoy.no' => true,
-    'bjarkøy.no' => true,
     'bjerkreim.no' => true,
     'bjugn.no' => true,
     'bodo.no' => true,
@@ -4227,7 +4245,6 @@ return array (
     'sande.møre-og-romsdal.no' => true,
     'moskenes.no' => true,
     'moss.no' => true,
-    'mosvik.no' => true,
     'muosat.no' => true,
     'muosát.no' => true,
     'naamesjevuemie.no' => true,
@@ -4996,6 +5013,7 @@ return array (
     'si' => true,
     'sj' => true,
     'sk' => true,
+    'org.sk' => true,
     'sl' => true,
     'com.sl' => true,
     'edu.sl' => true,
@@ -5009,7 +5027,6 @@ return array (
     'edu.sn' => true,
     'gouv.sn' => true,
     'org.sn' => true,
-    'perso.sn' => true,
     'univ.sn' => true,
     'so' => true,
     'com.so' => true,
@@ -5545,6 +5562,7 @@ return array (
     'firm.ve' => true,
     'gob.ve' => true,
     'gov.ve' => true,
+    'ia.ve' => true,
     'info.ve' => true,
     'int.ve' => true,
     'mil.ve' => true,
@@ -6065,7 +6083,6 @@ return array (
     'drive' => true,
     'dtv' => true,
     'dubai' => true,
-    'dunlop' => true,
     'dupont' => true,
     'durban' => true,
     'dvag' => true,
@@ -6905,9 +6922,6 @@ return array (
     'poznan.pl' => true,
     'wroc.pl' => true,
     'zakopane.pl' => true,
-    '12chars.dev' => true,
-    '12chars.it' => true,
-    '12chars.pro' => true,
     'cc.ua' => true,
     'inf.ua' => true,
     'ltd.ua' => true,
@@ -7028,6 +7042,7 @@ return array (
     'auth-fips.us-west-1.amazoncognito.com' => true,
     'auth.us-west-2.amazoncognito.com' => true,
     'auth-fips.us-west-2.amazoncognito.com' => true,
+    'auth.cognito-idp.eusc-de-east-1.on.amazonwebservices.eu' => true,
     'us-east-1.amazonaws.com' => true,
     'emrappui-prod.cn-north-1.amazonaws.com.cn' => true,
     'emrnotebooks-prod.cn-north-1.amazonaws.com.cn' => true,
@@ -7393,6 +7408,7 @@ return array (
     's3-accesspoint.dualstack.us-gov-east-1.amazonaws.com' => true,
     's3-accesspoint-fips.dualstack.us-gov-east-1.amazonaws.com' => true,
     's3-fips.dualstack.us-gov-east-1.amazonaws.com' => true,
+    's3-website.dualstack.us-gov-east-1.amazonaws.com' => true,
     's3.us-gov-east-1.amazonaws.com' => true,
     's3-accesspoint.us-gov-east-1.amazonaws.com' => true,
     's3-accesspoint-fips.us-gov-east-1.amazonaws.com' => true,
@@ -7403,6 +7419,7 @@ return array (
     's3-accesspoint.dualstack.us-gov-west-1.amazonaws.com' => true,
     's3-accesspoint-fips.dualstack.us-gov-west-1.amazonaws.com' => true,
     's3-fips.dualstack.us-gov-west-1.amazonaws.com' => true,
+    's3-website.dualstack.us-gov-west-1.amazonaws.com' => true,
     's3.us-gov-west-1.amazonaws.com' => true,
     's3-accesspoint.us-gov-west-1.amazonaws.com' => true,
     's3-accesspoint-fips.us-gov-west-1.amazonaws.com' => true,
@@ -7606,14 +7623,18 @@ return array (
     'ap-southeast-1.elasticbeanstalk.com' => true,
     'ap-southeast-2.elasticbeanstalk.com' => true,
     'ap-southeast-3.elasticbeanstalk.com' => true,
+    'ap-southeast-5.elasticbeanstalk.com' => true,
+    'ap-southeast-7.elasticbeanstalk.com' => true,
     'ca-central-1.elasticbeanstalk.com' => true,
     'eu-central-1.elasticbeanstalk.com' => true,
     'eu-north-1.elasticbeanstalk.com' => true,
     'eu-south-1.elasticbeanstalk.com' => true,
+    'eu-south-2.elasticbeanstalk.com' => true,
     'eu-west-1.elasticbeanstalk.com' => true,
     'eu-west-2.elasticbeanstalk.com' => true,
     'eu-west-3.elasticbeanstalk.com' => true,
     'il-central-1.elasticbeanstalk.com' => true,
+    'me-central-1.elasticbeanstalk.com' => true,
     'me-south-1.elasticbeanstalk.com' => true,
     'sa-east-1.elasticbeanstalk.com' => true,
     'us-east-1.elasticbeanstalk.com' => true,
@@ -7623,6 +7644,28 @@ return array (
     'us-west-1.elasticbeanstalk.com' => true,
     'us-west-2.elasticbeanstalk.com' => true,
     'awsglobalaccelerator.com' => true,
+    'lambda-url.af-south-1.on.aws' => true,
+    'lambda-url.ap-east-1.on.aws' => true,
+    'lambda-url.ap-northeast-1.on.aws' => true,
+    'lambda-url.ap-northeast-2.on.aws' => true,
+    'lambda-url.ap-northeast-3.on.aws' => true,
+    'lambda-url.ap-south-1.on.aws' => true,
+    'lambda-url.ap-southeast-1.on.aws' => true,
+    'lambda-url.ap-southeast-2.on.aws' => true,
+    'lambda-url.ap-southeast-3.on.aws' => true,
+    'lambda-url.ca-central-1.on.aws' => true,
+    'lambda-url.eu-central-1.on.aws' => true,
+    'lambda-url.eu-north-1.on.aws' => true,
+    'lambda-url.eu-south-1.on.aws' => true,
+    'lambda-url.eu-west-1.on.aws' => true,
+    'lambda-url.eu-west-2.on.aws' => true,
+    'lambda-url.eu-west-3.on.aws' => true,
+    'lambda-url.me-south-1.on.aws' => true,
+    'lambda-url.sa-east-1.on.aws' => true,
+    'lambda-url.us-east-1.on.aws' => true,
+    'lambda-url.us-east-2.on.aws' => true,
+    'lambda-url.us-west-1.on.aws' => true,
+    'lambda-url.us-west-2.on.aws' => true,
     'transfer-webapp.af-south-1.on.aws' => true,
     'transfer-webapp.ap-east-1.on.aws' => true,
     'transfer-webapp.ap-northeast-1.on.aws' => true,
@@ -7635,6 +7678,7 @@ return array (
     'transfer-webapp.ap-southeast-3.on.aws' => true,
     'transfer-webapp.ap-southeast-4.on.aws' => true,
     'transfer-webapp.ap-southeast-5.on.aws' => true,
+    'transfer-webapp.ap-southeast-7.on.aws' => true,
     'transfer-webapp.ca-central-1.on.aws' => true,
     'transfer-webapp.ca-west-1.on.aws' => true,
     'transfer-webapp.eu-central-1.on.aws' => true,
@@ -7648,6 +7692,7 @@ return array (
     'transfer-webapp.il-central-1.on.aws' => true,
     'transfer-webapp.me-central-1.on.aws' => true,
     'transfer-webapp.me-south-1.on.aws' => true,
+    'transfer-webapp.mx-central-1.on.aws' => true,
     'transfer-webapp.sa-east-1.on.aws' => true,
     'transfer-webapp.us-east-1.on.aws' => true,
     'transfer-webapp.us-east-2.on.aws' => true,
@@ -7661,13 +7706,16 @@ return array (
     'transfer-webapp.cn-northwest-1.on.amazonwebservices.com.cn' => true,
     'eero.online' => true,
     'eero-stage.online' => true,
+    'antagonist.cloud' => true,
     'apigee.io' => true,
     'panel.dev' => true,
     'siiites.com' => true,
+    'int.apple' => true,
     'appspacehosted.com' => true,
     'appspaceusercontent.com' => true,
     'appudo.net' => true,
     'appwrite.global' => true,
+    'appwrite.network' => true,
     'on-aptible.com' => true,
     'f5.si' => true,
     'arvanedge.ir' => true,
@@ -7681,10 +7729,8 @@ return array (
     'sweetpepper.org' => true,
     'myasustor.com' => true,
     'cdn.prod.atlassian-dev.net' => true,
-    'translated.page' => true,
     'myfritz.link' => true,
     'myfritz.net' => true,
-    'onavstack.net' => true,
     'ecommerce-shop.pl' => true,
     'b-data.io' => true,
     'balena-devices.com' => true,
@@ -7699,6 +7745,7 @@ return array (
     'shopselect.net' => true,
     'base.shop' => true,
     'beagleboard.io' => true,
+    'bearblog.dev' => true,
     'pages.gay' => true,
     'bnr.la' => true,
     'bitbucket.io' => true,
@@ -7716,6 +7763,7 @@ return array (
     'bplaced.net' => true,
     'square7.net' => true,
     'brave.app' => true,
+    'brave.dev' => true,
     'brave.io' => true,
     'shop.brendly.ba' => true,
     'shop.brendly.hr' => true,
@@ -7732,6 +7780,8 @@ return array (
     'canva-apps.cn' => true,
     'my.canvasite.cn' => true,
     'canva-apps.com' => true,
+    'canva-hosted-embed.com' => true,
+    'canvacode.com' => true,
     'rice-labs.com' => true,
     'canva.run' => true,
     'my.canva.site' => true,
@@ -7769,6 +7819,7 @@ return array (
     'ae.org' => true,
     'com.se' => true,
     'cx.ua' => true,
+    'discourse.diy' => true,
     'discourse.group' => true,
     'discourse.team' => true,
     'clerk.app' => true,
@@ -7811,6 +7862,7 @@ return array (
     'freesite.host' => true,
     'cloudaccess.net' => true,
     'cloudbeesusercontent.io' => true,
+    'cloudflare.app' => true,
     'cf-ipfs.com' => true,
     'cloudflare-ipfs.com' => true,
     'trycloudflare.com' => true,
@@ -7837,6 +7889,7 @@ return array (
     'co.nl' => true,
     'co.no' => true,
     'webhosting.be' => true,
+    'prvw.eu' => true,
     'hosting-cluster.nl' => true,
     'ctfcloud.net' => true,
     'convex.app' => true,
@@ -7847,6 +7900,7 @@ return array (
     'gov.ru' => true,
     'int.ru' => true,
     'mil.ru' => true,
+    'corespeed.app' => true,
     'dyn.cosidns.de' => true,
     'dnsupdater.de' => true,
     'dynamisches-dns.de' => true,
@@ -7897,9 +7951,16 @@ return array (
     'dedyn.io' => true,
     'deta.app' => true,
     'deta.dev' => true,
+    'deuxfleurs.eu' => true,
+    'deuxfleurs.page' => true,
+    'd6.ply.gg' => true,
+    'joinmc.link' => true,
+    'playit.plus' => true,
+    'with.playit.plus' => true,
     'icp0.io' => true,
     'icp1.io' => true,
     'caffeine.site' => true,
+    'caffeine.xyz' => true,
     'dfirma.pl' => true,
     'dkonto.pl' => true,
     'you2.pl' => true,
@@ -7911,9 +7972,16 @@ return array (
     'discordsays.com' => true,
     'discordsez.com' => true,
     'jozi.biz' => true,
+    'ccwu.cc' => true,
+    'cc.cd' => true,
+    'us.ci' => true,
+    'de5.net' => true,
     'dnshome.de' => true,
     'online.th' => true,
     'shop.th' => true,
+    'co.scot' => true,
+    'me.scot' => true,
+    'org.scot' => true,
     'drayddns.com' => true,
     'shoparena.pl' => true,
     'dreamhosters.com' => true,
@@ -8200,20 +8268,32 @@ return array (
     'stuff-4-sale.us' => true,
     'dyndns.ws' => true,
     'mypets.ws' => true,
+    '1cooldns.com' => true,
+    'bumbleshrimp.com' => true,
     'ddnsfree.com' => true,
     'ddnsgeek.com' => true,
+    'ddnsguru.com' => true,
+    'dynuddns.com' => true,
+    'dynuhosting.com' => true,
     'giize.com' => true,
     'gleeze.com' => true,
     'kozow.com' => true,
     'loseyourip.com' => true,
     'ooguy.com' => true,
+    'pivohosting.com' => true,
     'theworkpc.com' => true,
+    'wiredbladehosting.com' => true,
     'casacam.net' => true,
     'dynu.net' => true,
+    'dynuddns.net' => true,
+    'mysynology.net' => true,
+    'opik.net' => true,
+    'spryt.net' => true,
     'accesscam.org' => true,
     'camdvr.org' => true,
     'freeddns.org' => true,
     'mywire.org' => true,
+    'roxa.org' => true,
     'webredirect.org' => true,
     'myddns.rocks' => true,
     'dynv6.net' => true,
@@ -8229,7 +8309,9 @@ return array (
     'rt.ht' => true,
     'elementor.cloud' => true,
     'elementor.cool' => true,
-    'en-root.fr' => true,
+    'emergent.cloud' => true,
+    'preview.emergentagent.com' => true,
+    'emergent.host' => true,
     'mytuleap.com' => true,
     'tuleap-partners.com' => true,
     'encr.app' => true,
@@ -8408,17 +8490,15 @@ return array (
     'mydobiss.com' => true,
     'fh-muenster.io' => true,
     'figma.site' => true,
+    'figma-gov.site' => true,
     'preview.site' => true,
     'filegear.me' => true,
     'firebaseapp.com' => true,
     'fldrv.com' => true,
     'on-fleek.app' => true,
     'flutterflow.app' => true,
+    'sprites.app' => true,
     'fly.dev' => true,
-    'shw.io' => true,
-    'edgeapp.net' => true,
-    'forgeblocks.com' => true,
-    'id.forgerock.io' => true,
     'e2b.app' => true,
     'framer.ai' => true,
     'framer.app' => true,
@@ -8453,6 +8533,8 @@ return array (
     'mydns.vc' => true,
     'futurehosting.at' => true,
     'futuremailing.at' => true,
+    'gadget.app' => true,
+    'gadget.host' => true,
     'aliases121.com' => true,
     'campaign.gov.uk' => true,
     'service.gov.uk' => true,
@@ -8469,6 +8551,7 @@ return array (
     'gentlentapis.com' => true,
     'cdn-edges.net' => true,
     'gsj.bz' => true,
+    'gitbook.io' => true,
     'github.app' => true,
     'githubusercontent.com' => true,
     'githubpreview.dev' => true,
@@ -8607,8 +8690,10 @@ return array (
     'gov.nl' => true,
     'grafana-dev.net' => true,
     'grayjayleagues.com' => true,
+    'grebedoc.dev' => true,
     'günstigbestellen.de' => true,
     'günstigliefern.de' => true,
+    'gv.uy' => true,
     'hackclub.app' => true,
     'häkkinen.fi' => true,
     'hashbang.sh' => true,
@@ -8626,6 +8711,9 @@ return array (
     'helioho.st' => true,
     'heliohost.us' => true,
     'hepforge.org' => true,
+    'onhercules.app' => true,
+    'hercules-app.com' => true,
+    'hercules-dev.com' => true,
     'herokuapp.com' => true,
     'heyflow.page' => true,
     'heyflow.site' => true,
@@ -8661,6 +8749,7 @@ return array (
     'iki.fi' => true,
     'ibxos.it' => true,
     'iliadboxos.it' => true,
+    'imagine-proxy.work' => true,
     'smushcdn.com' => true,
     'wphostedmail.com' => true,
     'wpmucdn.com' => true,
@@ -8725,7 +8814,6 @@ return array (
     'ipifony.net' => true,
     'ir.md' => true,
     'is-a-good.dev' => true,
-    'is-a.dev' => true,
     'iservschule.de' => true,
     'mein-iserv.de' => true,
     'schuldock.de' => true,
@@ -8734,6 +8822,7 @@ return array (
     'test-iserv.de' => true,
     'iserv.dev' => true,
     'iserv.host' => true,
+    'ispmanager.name' => true,
     'mel.cloudlets.com.au' => true,
     'cloud.interhostsolutions.be' => true,
     'alp1.ae.flow.ch' => true,
@@ -8807,8 +8896,9 @@ return array (
     'myjino.ru' => true,
     'jote.cloud' => true,
     'jotelulu.cloud' => true,
-    'jote-dr-lt1.com' => true,
-    'jote-rd-lt1.com' => true,
+    'eu1-plenit.com' => true,
+    'la1-plenit.com' => true,
+    'us1-plenit.com' => true,
     'webadorsite.com' => true,
     'jouwweb.site' => true,
     'js.org' => true,
@@ -8817,7 +8907,14 @@ return array (
     'kapsi.fi' => true,
     'ezproxy.kuleuven.be' => true,
     'kuleuven.cloud' => true,
+    'keenetic.io' => true,
+    'keenetic.link' => true,
+    'keenetic.name' => true,
+    'keenetic.pro' => true,
+    'ae.kg' => true,
     'keymachine.de' => true,
+    'kiloapps.ai' => true,
+    'kiloapps.io' => true,
     'kinghost.net' => true,
     'uni5.net' => true,
     'knightpoint.systems' => true,
@@ -8826,6 +8923,8 @@ return array (
     'krellian.net' => true,
     'oya.to' => true,
     'co.de' => true,
+    'shiptoday.app' => true,
+    'shiptoday.build' => true,
     'laravel.cloud' => true,
     'on-forge.com' => true,
     'on-vapor.com' => true,
@@ -8856,7 +8955,6 @@ return array (
     'we.bs' => true,
     'filegear-sg.me' => true,
     'ggff.net' => true,
-    'localcert.net' => true,
     'localtonet.com' => true,
     'lodz.pl' => true,
     'pabianice.pl' => true,
@@ -8915,6 +9013,8 @@ return array (
     'barsyonline.co.uk' => true,
     'luyani.app' => true,
     'luyani.net' => true,
+    'magicpatterns.app' => true,
+    'magicpatternsapp.com' => true,
     'hb.cldmail.ru' => true,
     'matlab.cloud' => true,
     'modelscape.com' => true,
@@ -8962,8 +9062,18 @@ return array (
     'azurewebsites.net' => true,
     'cloudapp.net' => true,
     'trafficmanager.net' => true,
+    'blob.core.usgovcloudapi.net' => true,
+    'file.core.usgovcloudapi.net' => true,
+    'web.core.usgovcloudapi.net' => true,
+    'servicebus.usgovcloudapi.net' => true,
+    'usgovcloudapp.net' => true,
+    'usgovtrafficmanager.net' => true,
     'blob.core.windows.net' => true,
+    'file.core.windows.net' => true,
+    'web.core.windows.net' => true,
     'servicebus.windows.net' => true,
+    'azure-api.us' => true,
+    'azurewebsites.us' => true,
     'routingthecloud.com' => true,
     'sn.mynetname.net' => true,
     'routingthecloud.net' => true,
@@ -8971,12 +9081,17 @@ return array (
     'same-app.com' => true,
     'same-preview.com' => true,
     'csx.cc' => true,
+    'miren.app' => true,
+    'miren.systems' => true,
     'mydbserver.com' => true,
     'webspaceconfig.de' => true,
     'mittwald.info' => true,
     'mittwaldserver.info' => true,
     'typo3server.info' => true,
     'project.space' => true,
+    'mocha.app' => true,
+    'mochausercontent.com' => true,
+    'mocha-sandbox.dev' => true,
     'modx.dev' => true,
     'bmoattachments.org' => true,
     'net.ru' => true,
@@ -8996,10 +9111,11 @@ return array (
     'yali.mythic-beasts.com' => true,
     'cust.retrosnub.co.uk' => true,
     'ui.nabu.casa' => true,
+    'needle.run' => true,
+    'co.site' => true,
     'cloud.nospamproxy.com' => true,
     'o365.cloud.nospamproxy.com' => true,
     'netlib.re' => true,
-    'netfy.app' => true,
     'netlify.app' => true,
     '4u.com' => true,
     'nfshost.com' => true,
@@ -9129,9 +9245,7 @@ return array (
     'freeddns.us' => true,
     'nsupdate.info' => true,
     'nerdpol.ovh' => true,
-    'nyc.mn' => true,
     'prvcy.page' => true,
-    'obl.ong' => true,
     'observablehq.cloud' => true,
     'static.observableusercontent.com' => true,
     'omg.lol' => true,
@@ -9153,6 +9267,7 @@ return array (
     '123website.nl' => true,
     '123hjemmeside.no' => true,
     'service.one' => true,
+    'website.one' => true,
     'simplesite.pl' => true,
     '123paginaweb.pt' => true,
     '123minsida.se' => true,
@@ -9173,7 +9288,6 @@ return array (
     'can.re' => true,
     'authgear-staging.com' => true,
     'authgearapps.com' => true,
-    'skygearapp.com' => true,
     'outsystemscloud.com' => true,
     'ownprovider.com' => true,
     'own.pm' => true,
@@ -9188,17 +9302,16 @@ return array (
     'srv.us' => true,
     'gh.srv.us' => true,
     'gl.srv.us' => true,
-    'lk3.ru' => true,
     'mypep.link' => true,
     'perspecta.cloud' => true,
+    'forgeblocks.com' => true,
+    'id.forgerock.io' => true,
     'support.site' => true,
     'on-web.fr' => true,
     'upsunapp.com' => true,
     'ent.platform.sh' => true,
     'eu.platform.sh' => true,
     'us.platform.sh' => true,
-    'platter-app.dev' => true,
-    'platterp.us' => true,
     'pley.games' => true,
     'onporter.run' => true,
     'co.bn' => true,
@@ -9210,15 +9323,18 @@ return array (
     'xen.prgmr.com' => true,
     'priv.at' => true,
     'c01.kr' => true,
+    'eliv-api.kr' => true,
     'eliv-cdn.kr' => true,
     'eliv-dns.kr' => true,
     'mmv.kr' => true,
     'vki.kr' => true,
     'dev.project-study.com' => true,
     'protonet.io' => true,
+    'platter-app.dev' => true,
     'e.id' => true,
     'chirurgiens-dentistes-en-france.fr' => true,
     'byen.site' => true,
+    'nyc.mn' => true,
     'pubtls.org' => true,
     'pythonanywhere.com' => true,
     'eu.pythonanywhere.com' => true,
@@ -9361,6 +9477,7 @@ return array (
     'from.tv' => true,
     'sakura.tv' => true,
     'sandcats.io' => true,
+    'sav.case' => true,
     'logoip.com' => true,
     'logoip.de' => true,
     'fr-par-1.baremetal.scw.cloud' => true,
@@ -9465,6 +9582,7 @@ return array (
     'streamlitapp.com' => true,
     'try-snowplow.com' => true,
     'mafelo.net' => true,
+    'sol.site' => true,
     'playstation-cloud.com' => true,
     'srht.site' => true,
     'apps.lair.io' => true,
@@ -9472,6 +9590,7 @@ return array (
     'my.at' => true,
     'my.de' => true,
     'nx.gw' => true,
+    'spawnbase.app' => true,
     'customer.speedpartner.de' => true,
     'myspreadshop.at' => true,
     'myspreadshop.com.au' => true,
@@ -9502,9 +9621,11 @@ return array (
     'stackit.rocks' => true,
     'stackit.run' => true,
     'stackit.zone' => true,
+    'indevs.in' => true,
     'musician.io' => true,
     'novecore.site' => true,
     'api.stdlib.com' => true,
+    'statichost.page' => true,
     'feedback.ac' => true,
     'forms.ac' => true,
     'assessments.cx' => true,
@@ -9518,7 +9639,6 @@ return array (
     'ipfs.storacha.link' => true,
     'ipfs.w3s.link' => true,
     'storebase.store' => true,
-    'storipress.app' => true,
     'storj.farm' => true,
     'strapiapp.com' => true,
     'media.strapiapp.com' => true,
@@ -9572,6 +9692,12 @@ return array (
     'p.tawkto.email' => true,
     'tche.br' => true,
     'site.tb-hosting.com' => true,
+    'directwp.eu' => true,
+    'ec.cc' => true,
+    'eu.cc' => true,
+    'gu.cc' => true,
+    'uk.cc' => true,
+    'us.cc' => true,
     'edugit.io' => true,
     's3.teckids.org' => true,
     'telebit.app' => true,
@@ -9613,20 +9739,8 @@ return array (
     'webspace.rocks' => true,
     'lima.zone' => true,
     'site.transip.me' => true,
+    'tunnelmole.net' => true,
     'tuxfamily.org' => true,
-    'dd-dns.de' => true,
-    'dray-dns.de' => true,
-    'draydns.de' => true,
-    'dyn-vpn.de' => true,
-    'dynvpn.de' => true,
-    'mein-vigor.de' => true,
-    'my-vigor.de' => true,
-    'my-wan.de' => true,
-    'syno-ds.de' => true,
-    'synology-diskstation.de' => true,
-    'synology-ds.de' => true,
-    'diskstation.eu' => true,
-    'diskstation.org' => true,
     'typedream.app' => true,
     'pro.typeform.com' => true,
     'uber.space' => true,
@@ -9635,6 +9749,7 @@ return array (
     'ltd.hk' => true,
     'hk.org' => true,
     'it.com' => true,
+    'umso.co' => true,
     'unison-services.cloud' => true,
     'virtual-user.de' => true,
     'virtualuser.de' => true,
@@ -9665,6 +9780,7 @@ return array (
     'voorloper.cloud' => true,
     'wafflecell.com' => true,
     'wal.app' => true,
+    'wasmer.app' => true,
     'webflow.io' => true,
     'webflowtest.io' => true,
     'bookonline.app' => true,
@@ -9685,14 +9801,16 @@ return array (
     'wmcloud.org' => true,
     'beta.wmcloud.org' => true,
     'wmflabs.org' => true,
-    'wdh.app' => true,
-    'hrsn.au' => true,
     'vps.hrsn.au' => true,
     'hrsn.dev' => true,
+    'is-a.dev' => true,
+    'localcert.net' => true,
     'windsurf.app' => true,
     'windsurf.build' => true,
     'panel.gg' => true,
     'daemon.panel.gg' => true,
+    'base44.app' => true,
+    'base44-sandbox.com' => true,
     'wixsite.com' => true,
     'wixstudio.com' => true,
     'editorx.io' => true,
@@ -9717,6 +9835,7 @@ return array (
     'cistron.nl' => true,
     'demon.nl' => true,
     'xs4all.space' => true,
+    'xtooldevice.com' => true,
     'yandexcloud.net' => true,
     'storage.yandexcloud.net' => true,
     'website.yandexcloud.net' => true,
@@ -9735,11 +9854,11 @@ return array (
     'virtualserver.io' => true,
     'enterprisecloud.nu' => true,
     'zone.id' => true,
+    'nett.to' => true,
     'zabc.net' => true,
   ),
   'WILDCARD' => 
   array (
-    'bd' => true,
     'nom.br' => true,
     'ck' => true,
     'er' => true,
@@ -9759,6 +9878,7 @@ return array (
     'sch.uk' => true,
     'on-acorn.io' => true,
     'dev.adobeaemcloud.com' => true,
+    'auiusercontent.com' => true,
     'compute.estate' => true,
     'alces.network' => true,
     'compute.amazonaws.com.cn' => true,
@@ -9810,6 +9930,7 @@ return array (
     'ap-southeast-3.airflow.amazonaws.com' => true,
     'ap-southeast-4.airflow.amazonaws.com' => true,
     'ap-southeast-5.airflow.amazonaws.com' => true,
+    'ap-southeast-7.airflow.amazonaws.com' => true,
     'ca-central-1.airflow.amazonaws.com' => true,
     'ca-west-1.airflow.amazonaws.com' => true,
     'eu-central-1.airflow.amazonaws.com' => true,
@@ -9828,17 +9949,69 @@ return array (
     'us-east-2.airflow.amazonaws.com' => true,
     'us-west-1.airflow.amazonaws.com' => true,
     'us-west-2.airflow.amazonaws.com' => true,
+    'rds.cn-north-1.amazonaws.com.cn' => true,
+    'rds.cn-northwest-1.amazonaws.com.cn' => true,
+    'af-south-1.rds.amazonaws.com' => true,
+    'ap-east-1.rds.amazonaws.com' => true,
+    'ap-east-2.rds.amazonaws.com' => true,
+    'ap-northeast-1.rds.amazonaws.com' => true,
+    'ap-northeast-2.rds.amazonaws.com' => true,
+    'ap-northeast-3.rds.amazonaws.com' => true,
+    'ap-south-1.rds.amazonaws.com' => true,
+    'ap-south-2.rds.amazonaws.com' => true,
+    'ap-southeast-1.rds.amazonaws.com' => true,
+    'ap-southeast-2.rds.amazonaws.com' => true,
+    'ap-southeast-3.rds.amazonaws.com' => true,
+    'ap-southeast-4.rds.amazonaws.com' => true,
+    'ap-southeast-5.rds.amazonaws.com' => true,
+    'ap-southeast-6.rds.amazonaws.com' => true,
+    'ap-southeast-7.rds.amazonaws.com' => true,
+    'ca-central-1.rds.amazonaws.com' => true,
+    'ca-west-1.rds.amazonaws.com' => true,
+    'eu-central-1.rds.amazonaws.com' => true,
+    'eu-central-2.rds.amazonaws.com' => true,
+    'eu-west-1.rds.amazonaws.com' => true,
+    'eu-west-2.rds.amazonaws.com' => true,
+    'eu-west-3.rds.amazonaws.com' => true,
+    'il-central-1.rds.amazonaws.com' => true,
+    'me-central-1.rds.amazonaws.com' => true,
+    'me-south-1.rds.amazonaws.com' => true,
+    'mx-central-1.rds.amazonaws.com' => true,
+    'sa-east-1.rds.amazonaws.com' => true,
+    'us-east-1.rds.amazonaws.com' => true,
+    'us-east-2.rds.amazonaws.com' => true,
+    'us-gov-east-1.rds.amazonaws.com' => true,
+    'us-gov-west-1.rds.amazonaws.com' => true,
+    'us-northeast-1.rds.amazonaws.com' => true,
+    'us-west-1.rds.amazonaws.com' => true,
+    'us-west-2.rds.amazonaws.com' => true,
     'experiments.sagemaker.aws' => true,
     'awsapprunner.com' => true,
     'elb.amazonaws.com.cn' => true,
     'elb.amazonaws.com' => true,
     'private.repost.aws' => true,
+    'cloud.int.apple' => true,
+    'r.cloud.int.apple' => true,
+    'ap-north-1.r.cloud.int.apple' => true,
+    'ap-south-1.r.cloud.int.apple' => true,
+    'ap-south-2.r.cloud.int.apple' => true,
+    'eu-central-1.r.cloud.int.apple' => true,
+    'eu-north-1.r.cloud.int.apple' => true,
+    'us-central-1.r.cloud.int.apple' => true,
+    'us-central-2.r.cloud.int.apple' => true,
+    'us-east-1.r.cloud.int.apple' => true,
+    'us-east-2.r.cloud.int.apple' => true,
+    'us-west-1.r.cloud.int.apple' => true,
+    'us-west-2.r.cloud.int.apple' => true,
+    'us-west-3.r.cloud.int.apple' => true,
     'appwrite.run' => true,
     'awdev.ca' => true,
     'advisor.ws' => true,
     'beget.app' => true,
     's.brave.app' => true,
+    's.brave.dev' => true,
     's.brave.io' => true,
+    'bwcloud-os-instance.de' => true,
     'lcl.dev' => true,
     'lclstage.dev' => true,
     'stg.dev' => true,
@@ -9848,6 +10021,8 @@ return array (
     'otap.co' => true,
     'devinapps.com' => true,
     'cryptonomic.net' => true,
+    'at.ply.gg' => true,
+    'at.playit.plus' => true,
     'raw.icp0.io' => true,
     'raw.icp1.io' => true,
     'icp.net' => true,
@@ -9880,7 +10055,6 @@ return array (
     'landing.myjino.ru' => true,
     'spectrum.myjino.ru' => true,
     'vps.myjino.ru' => true,
-    'cns.joyent.com' => true,
     'triton.zone' => true,
     'nodebalancer.linode.com' => true,
     'linodeobjects.com' => true,
@@ -9912,6 +10086,7 @@ return array (
     'upsun.app' => true,
     'platformsh.site' => true,
     'tst.site' => true,
+    'cn.st' => true,
     'sys.qcx.io' => true,
     'qualyhqpartner.com' => true,
     'qualyhqportal.com' => true,
@@ -9919,11 +10094,26 @@ return array (
     'on-rancher.cloud' => true,
     'on-k3s.io' => true,
     'on-rio.io' => true,
+    'clusters.rdpa.co' => true,
+    'srvrless.rdpa.co' => true,
     'builder.code.com' => true,
     'dev-builder.code.com' => true,
     'stg-builder.code.com' => true,
     '001.test.code-builder-stg.platform.salesforce.com' => true,
+    'aa.crm.dev' => true,
+    'ab.crm.dev' => true,
+    'ac.crm.dev' => true,
+    'ad.crm.dev' => true,
+    'ae.crm.dev' => true,
+    'af.crm.dev' => true,
+    'ci.crm.dev' => true,
     'd.crm.dev' => true,
+    'pa.crm.dev' => true,
+    'pb.crm.dev' => true,
+    'pc.crm.dev' => true,
+    'pd.crm.dev' => true,
+    'pe.crm.dev' => true,
+    'pf.crm.dev' => true,
     'w.crm.dev' => true,
     'wa.crm.dev' => true,
     'wb.crm.dev' => true,
